+++ lib/tabulatr/renderer/renderer.rb
@@ -32,7 +32,8 @@ class Tabulatr::Renderer
       batch_actions: false,  # :name => value hash of batch action stuff
       footer_content: false, # if given, add a <%= content_for <footer_content> %> before the </table>
       path: '#',             # where to send the AJAX-requests to
-      order_by: nil)         # default order
+      order_by: nil,
+      html_class: '')         # default order
     @klass = klass
     @view = view
     @table_options = {
@@ -44,7 +45,8 @@ class Tabulatr::Renderer
       batch_actions: batch_actions,
       footer_content: footer_content,
       path: path,
-      order_by: order_by
+      order_by: order_by,
+      html_class: 'table tabulatr_table '.concat(html_class)
     }
     @classname = @klass.name.underscore
   end